+++ tests/fixtures/noqa/noqa.py
@@ -767,3 +767,9 @@ def get_item():  # noqa: WPS463
     return  # noqa: WPS324
 
 bad_bitwise = True | True # noqa: WPS465
+
+matrix = [ # noqa: WPS361
+   some(number) for numbers in matrix # noqa: WPS361
+   for number in numbers # noqa: WPS361
+   if number > 0 # noqa: WPS361
+] # noqa: WPS361