+++ tests/test_checker/test_noqa.py
@@ -171,6 +171,7 @@ SHOULD_BE_RAISED = types.MappingProxyType({
     'WPS358': 1,
     'WPS359': 1,
     'WPS360': 1,
+    'WPS361': 1,
 
     'WPS400': 0,  # defined in ignored violations.
     'WPS401': 0,  # logically unacceptable.