+++ wemake_python_styleguide/visitors/tokenize/statements.py
@@ -1,5 +1,6 @@
 import tokenize
 from collections import defaultdict
+from dataclasses import dataclass
 from operator import attrgetter
 from typing import (
     ClassVar,
@@ -30,9 +31,11 @@ from wemake_python_styleguide.violations.best_practices import (
 from wemake_python_styleguide.violations.consistency import (
     BracketBlankLineViolation,
     ExtraIndentationViolation,
+    InconsistentComprehensionViolation,
     WrongBracketPositionViolation,
 )
 from wemake_python_styleguide.visitors.base import BaseTokenVisitor
+from wemake_python_styleguide.visitors.decorators import alias
 
 TokenLines = DefaultDict[int, List[tokenize.TokenInfo]]
 
@@ -256,3 +259,197 @@ class MultilineStringVisitor(BaseTokenVisitor):
             self._check_individual_line(
                 line_tokens, previous_line_token, next_line_token,
             )
+
+
+@final
+@dataclass
+class InconsistentComprehensionContext(object):
+    r"""Context for individual bracket enclosures (i.e. [],\{\}, or ())."""
+
+    def __init__(self) -> None:
+        r"""
+        Initializes context for a given bracket enclosure (i.e. [],\{\}, or ()).
+
+        Helper class for InconsistentComprehensionVisitor which stores context
+        for the current (potential) comprehension we are in. Combined with a
+        stack to enable support for nested comprehensions.
+
+        _is_comprehension:
+        Flag is set if current clause is identified as a list comprehension.
+
+        seen_clause_in_line:
+        Flag is set when the current line already contains a clause, which
+        is either the action, each for loop, or the conditional. Starts off
+        as True to account for the action, which we don't actually visit.
+
+        _seen_for:
+        Flag tracks whether we've seen a for statement within these brackets.
+        Effectively determines whether we are looking at some kind of
+        comprehension or not.
+
+        _seen_for_in_line:
+        Flag tracks whether we've seen a for statement on this line. Used to
+        determine if a for...in statement has been split across multiple lines.
+
+        _seen_if_in_line:
+        Flag tracks whether we've seen an if statement on this line. Used to
+        determine whether an in statement is from a for...in statement or
+        is a logical in.
+
+        _seen_nl:
+        Flag for if we've seen any logical newlines, indicating this is a
+        multiline comprehension
+
+        _potential_violation:
+        Flag for when we see multiple clauses in one line. Only a violation
+        if this is a multiline comprehension
+
+        _reported:
+        Flag tracks whether we've already reported this violation.
+        """
+        self.seen_clause_in_line = False
+        self._is_comprehension = False
+        self._seen_for = False
+        self._seen_for_in_line = False
+        self._seen_if_in_line = False
+        self._seen_nl = False
+        self._potential_violation = False
+        self._reported = False
+
+    def check_nl(self, token: tokenize.TokenInfo) -> None:
+        """
+        Handles logical newline character depending on context.
+
+        Sets appropriate flags to True if nl encountered inside brackets after
+        some clause, so that a single line comprehension with brackets on
+        multiple lines is still accepted.
+        """
+        self._seen_nl = self.seen_clause_in_line
+        self.seen_clause_in_line = False
+        self._seen_for_in_line = False
+        self._seen_if_in_line = False
+
+    def check_for(self, token: tokenize.TokenInfo) -> bool:
+        """Handles 'for' tokens inside brackets."""
+        self._is_comprehension = True
+        self._seen_for = True
+        self._seen_for_in_line = True
+
+        self._potential_violation = (
+            self._potential_violation or
+            self.seen_clause_in_line
+        )
+        return self._check_violation(token)
+
+    def check_if(self, token: tokenize.TokenInfo) -> bool:
+        """
+        Handles 'if' tokens inside brackets.
+
+        In order to rule out this if statement being part of a ternary operator
+        in the action statement of a comprehension, we ensure that we've already
+        passed the action statement, as once a for statement appears any
+        following if statements will be the conditional statement of the
+        comprehension.
+        """
+        if self._seen_for:
+            self._seen_if_in_line = True
+
+            self._potential_violation = (
+                self._potential_violation or
+                self.seen_clause_in_line
+            )
+            return self._check_violation(token)
+        return True
+
+    def check_in(self, token: tokenize.TokenInfo) -> bool:
+        """
+        Adds a violation when a for...in statement is split across lines.
+
+        Because of the overloaded nature of 'in', we have to rule
+        out several edge cases before we can say that this is a violation:
+        First, we only want to consider this split for...in statement
+        in the case that we've seen a for statement but not on this line.
+        Second, we don't want to catch an 'in' in the conditional clause of the
+        comprehension, so we make sure we haven't seen an if statement on
+        this line.
+        """
+        if self._seen_for:
+            if not self._seen_for_in_line:
+                if not self._seen_if_in_line:
+                    self._reported = True
+                    return False
+        return True
+
+    def _check_violation(self, token: tokenize.TokenInfo) -> bool:
+        """Checks if current environment state implies a violation."""
+        if self._seen_nl:
+            if self._potential_violation and not self._reported:
+                self._reported = True
+                return False
+        return True
+
+
+@final
+@alias('visit_any_left_bracket', (
+    'visit_lsqb',
+    'visit_lbrace',
+    'visit_lpar',
+))
+@alias('visit_any_right_bracket', (
+    'visit_rsqb',
+    'visit_rbrace',
+    'visit_rpar',
+))
+class InconsistentComprehensionVisitor(BaseTokenVisitor):
+    """
+    Visitor for checking inconsistent comprehension syntax.
+
+    Checks if comprehensions either use only one line or inserts a newline
+    for each clause (i.e. bracket, action, for loop(s), and conditional)
+    """
+
+    def __init__(self, *args, **kwargs) -> None:
+        """
+        Initializes stack of bracket contexts.
+
+        Creates an empty stack for bracket contexts to accomodate for nested
+        comprehensions.
+        """
+        super().__init__(*args, **kwargs)
+        self._bracket_stack: List[InconsistentComprehensionContext] = []
+        self._current_ctx: Optional[InconsistentComprehensionContext] = None
+
+    def visit_any_left_bracket(self, token: tokenize.TokenInfo) -> None:
+        """Sets self._inside_brackets to True if left bracket found."""
+        self._bracket_stack.append(InconsistentComprehensionContext())
+        self._current_ctx = self._bracket_stack[-1]
+
+    def visit_any_right_bracket(self, token: tokenize.TokenInfo) -> None:
+        """Resets environment if right bracket is encountered."""
+        self._bracket_stack.pop()
+        if self._bracket_stack:
+            self._current_ctx = self._bracket_stack[-1]
+        else:
+            self._current_ctx = None
+
+    def visit_nl(self, token: tokenize.TokenInfo) -> None:
+        """Handles logical newline character depending on context."""
+        if self._current_ctx:
+            self._current_ctx.check_nl(token)
+
+    def visit_name(self, token: tokenize.TokenInfo) -> None:
+        """Sets flags for comprehension keywords."""
+        if self._current_ctx:
+            well_formed = True
+            if token.string == 'for':
+                well_formed = self._current_ctx.check_for(token)
+            elif token.string == 'if':
+                well_formed = self._current_ctx.check_if(token)
+            elif token.string == 'in':
+                well_formed = self._current_ctx.check_in(token)
+
+            self._current_ctx.seen_clause_in_line = True
+            if not well_formed:
+                self.add_violation(
+                    InconsistentComprehensionViolation(token),
+                )